requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
appdirs=1.4.4=pypi_0
blinker=1.7.0=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.11.17=h56e8100_0
certifi=2022.12.7=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
expat=2.5.0=hd77b12b_0
filelock=3.9.0=pypi_0
flask=2.3.3=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.47.2=pypi_0
fsspec=2023.4.0=pypi_0
humanfriendly=10.0=pypi_0
idna=3.4=pypi_0
intel-openmp=2024.0.0=h57928b3_49840
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
label-studio-ml=2.0.0.dev0=dev_0
label-studio-tools=0.0.3=pypi_0
libabseil=20230802.1=cxx17_h63175ca_0
libblas=3.9.0=21_win64_mkl
libcblas=3.9.0=21_win64_mkl
libexpat=2.5.0=h63175ca_1
libffi=3.4.4=hd77b12b_0
libhwloc=2.9.3=default_haede6df_1009
libiconv=1.17=hcfcfb64_2
liblapack=3.9.0=21_win64_mkl
libprotobuf=4.24.4=hb8276f3_0
libsqlite=3.44.2=hcfcfb64_0
libxml2=2.12.4=hc3477c8_1
libzlib=1.2.13=hcfcfb64_5
lxml=5.1.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=pypi_0
mdurl=0.1.2=pypi_0
mkl=2024.0.0=h66d3029_49657
mpmath=1.2.1=pypi_0
networkx=3.0rc1=pypi_0
numpy=1.26.3=py312h8753938_0
onnx=1.15.0=py312ha26593b_0
onnxruntime-gpu=1.17.0=pypi_0
onnxsim=0.4.33=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.2.0=hcfcfb64_1
packaging=23.2=pypi_0
pandas=2.2.0=pypi_0
pillow=10.1.0=pypi_0
pip=23.3.1=py312haa95532_0
protobuf=4.24.4=py312hd11f8d2_0
psutil=5.9.8=pypi_0
pthreads-win32=2.9.1=hfa6e2cd_3
py-cpuinfo=9.0.0=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
pyreadline3=3.4.1=pypi_0
python=3.12.1=h2628c8c_1_cpython
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.1=pypi_0
python_abi=3.12=4_cp312
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
requests=2.31.0=pypi_0
rich=13.7.0=pypi_0
scipy=1.12.0=pypi_0
seaborn=0.13.1=pypi_0
setuptools=69.0.3=pypi_0
six=1.16.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
sympy=1.11.1=pypi_0
tbb=2021.11.0=h91493d7_0
thop=0.1.1-2209072238=pypi_0
tk=8.6.13=h5226925_1
torch=2.3.0.dev20240122+cu121=pypi_0
torchaudio=2.2.0.dev20240122+cu121=pypi_0
torchvision=0.18.0.dev20240122+cu121=pypi_0
tqdm=4.66.1=pypi_0
typing-extensions=4.8.0=pypi_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023.4=pypi_0
ucrt=10.0.22621.0=h57928b3_0
ultralytics=8.1.5=pypi_0
urllib3=1.26.13=pypi_0
vc=14.2=h21ff451_1
vc14_runtime=14.38.33130=h82b7239_18
vs2015_runtime=14.38.33130=hcb4865c_18
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py312haa95532_0
xz=5.4.5=h8cc25b3_0
